EW-179 Refactor invalidUser

+++ src/services/sync/strategies/TSP/SchoolChange.js
@@ -1,5 +1,5 @@
 const { getUsername } = require('./TSP');
-const { FileModel } = require('../../../fileStorage/model.js');
+const { FileModel } = require('../../../fileStorage/model');
 const { info: logInfo, error: logError } = require('../../../../logger');
 
 const getInvalidatedUuid = (uuid) => `${uuid}/invalid!`;
@@ -7,7 +7,8 @@ const getInvalidatedEmail = (email) => `${email}.invalid`;
 
 const invalidateUser = async (app, user) => {
 	const userService = app.service('usersModel');
-	const accountService = app.service('accountModel');
+	// const accountService = app.service('accountModel');
+	const accountService = app.service('nest-account-service');
 
 	const invalidatedUuid = getInvalidatedUuid(user.sourceOptions.tspUid);
 	const userChanges = {
@@ -18,19 +19,18 @@ const invalidateUser = async (app, user) => {
 	};
 	await userService.patch(user._id, userChanges);
 
-	const accountChanges = {
-		username: getUsername(userChanges),
-	};
-	await accountService.patch(null, accountChanges, { query: { userId: user._id } });
+	// const accountChanges = {
+	// 	username: getUsername(userChanges),
+	// };
+	// await accountService.patch(null, accountChanges, { query: { userId: user._id } });
+	const account = await accountService.findByUserId(user._id);
+	await accountService.updateUsername(account.id, getUsername(userChanges));
 };
 
 const deleteUser = (app, user) => {
 	const userService = app.service('usersModel');
 	const accountService = app.service('nest-account-service');
-	return Promise.all([
-		userService.remove({ _id: user._id }),
-		accountService.deleteByUserId(user._id.toString()),
-	]);
+	return Promise.all([userService.remove({ _id: user._id }), accountService.deleteByUserId(user._id.toString())]);
 };
 
 const grantAccessToPrivateFiles = async (app, oldUser, newUser) => {
